Use readBinary vault method
