+++ src/main.ts
@@ -40,8 +40,6 @@ import {
 import { randomUUID } from "crypto";
 import {
 	combinePayloads,
-	getFileContents,
-	getImageContent,
 	getWordContents,
 	isImage,
 	isOtherText,
@@ -305,6 +303,49 @@ export default class CloudAtlasPlugin extends Plugin {
 		};
 	};
 
+	getFileContents = async (
+		basePath: string,
+		path: string
+	): Promise<string | null> => {
+		const contents = await this.app.vault.readBinary(
+			this.app.vault.getAbstractFileByPath(`${basePath}/${path}`) as TFile
+		);
+		try {
+			return new TextDecoder("utf8", { fatal: true }).decode(contents);
+		} catch (e) {
+			console.debug(e);
+			return null;
+		}
+	};
+
+	getImageNodeContent = async (
+		basePath: string,
+		node: FileNode
+	): Promise<string> => {
+		return this.getImageContent(basePath, node.file);
+	};
+
+	getImageContent = async (
+		basePath: string,
+		path: string
+	): Promise<string> => {
+		const contents = await this.app.vault.readBinary(
+			this.app.vault.getAbstractFileByPath(`${basePath}/${path}`) as TFile
+		);
+		const buffedInput = Buffer.from(contents).toString("base64");
+
+		// use the file extension to determine the mime type
+		// can we use a case statement here?
+		if (path.endsWith(".jpg") || path.endsWith(".jpeg")) {
+			return `data:image/jpeg;base64,${buffedInput}`;
+		} else if (path.endsWith(".gif")) {
+			return `data:image/gif;base64,${buffedInput}`;
+		}
+
+		// default to png
+		return `data:image/png;base64,${buffedInput}`;
+	};
+
 	parseExclusionPatterns = (patterns: string[]): RegExp[] => {
 		return patterns.map((pattern) => new RegExp(pattern));
 	};
@@ -407,13 +448,13 @@ export default class CloudAtlasPlugin extends Plugin {
 				return await this.canvasOps(getFileByPath(path, this.app));
 			}
 			if (isImage(path)) {
-				return await getImageContent(basePath, path);
+				return await this.getImageContent(basePath, path);
 			}
 			if (isWord(path)) {
 				return await getWordContents(basePath, path);
 			}
 			if (isOtherText(path)) {
-				return getFileContents(basePath, path);
+				return await this.getFileContents(basePath, path);
 			}
 			return await this.readNote(path);
 		} catch (e) {
