+++ src/utils.ts
@@ -1,4 +1,3 @@
-import { readFileSync } from "fs";
 import WordExtractor from "word-extractor";
 import { AdditionalContext, Payload, User } from "./interfaces";
 import {
@@ -89,25 +88,6 @@ export function joinStrings(
 	return [first, second].filter((s) => s).join("\n");
 }
 
-export async function getImageContent(
-	basePath: string,
-	path: string
-): Promise<string> {
-	const contents = readFileSync(`${basePath}/${path}`);
-	const buffedInput = Buffer.from(contents).toString("base64");
-
-	// use the file extension to determine the mime type
-	// can we use a case statement here?
-	if (path.endsWith(".jpg") || path.endsWith(".jpeg")) {
-		return `data:image/jpeg;base64,${buffedInput}`;
-	} else if (path.endsWith(".gif")) {
-		return `data:image/gif;base64,${buffedInput}`;
-	}
-
-	// default to png
-	return `data:image/png;base64,${buffedInput}`;
-}
-
 export function isImage(path: string): boolean {
 	return (
 		path.endsWith(".png") ||
@@ -144,14 +124,4 @@ export async function getWordContents(
 	const extractor = new WordExtractor();
 	const extracted = await extractor.extract(`${basePath}/${path}`);
 	return extracted.getBody();
-}
-
-export function getFileContents(basePath: string, path: string): string | null {
-	const contents = readFileSync(`${basePath}/${path}`);
-	try {
-		return new TextDecoder("utf8", { fatal: true }).decode(contents);
-	} catch (e) {
-		console.debug(e);
-		return null;
-	}
 }
